fix: Avoid short-circuit when expressions access non-existent attributes
If a boolean expression arm in one of `any`, `all` or `none` blocks
uses a non-existent attribute, it returns a `nil` value. This cannot be
cast to a bool so we raise an error and short-circuit the evaluation of
the whole block. This patch fixes that issue.

Fixes #1055
