+++ internal/engine/engine_test.go
@@ -17,7 +17,6 @@ import (
 	"google.golang.org/protobuf/types/known/structpb"
 
 	enginev1 "github.com/cerbos/cerbos/api/genpb/cerbos/engine/v1"
-	policyv1 "github.com/cerbos/cerbos/api/genpb/cerbos/policy/v1"
 	privatev1 "github.com/cerbos/cerbos/api/genpb/cerbos/private/v1"
 	schemav1 "github.com/cerbos/cerbos/api/genpb/cerbos/schema/v1"
 	"github.com/cerbos/cerbos/internal/audit"
@@ -200,39 +199,6 @@ func mkEngine(tb testing.TB, p param) (*Engine, context.CancelFunc) {
 	require.NoError(tb, err)
 
 	return eng, cancelFunc
-}
-
-func TestSatisfiesCondition(t *testing.T) {
-	testCases := test.LoadTestCases(t, "cel_eval")
-
-	for _, tcase := range testCases {
-		tcase := tcase
-		t.Run(tcase.Name, func(t *testing.T) {
-			tc := readCELTestCase(t, tcase.Input)
-			cond, err := compile.Condition(&policyv1.Condition{Condition: &policyv1.Condition_Match{Match: tc.Condition}})
-			require.NoError(t, err)
-
-			tctx := tracer.Start(newTestTraceSink(t))
-			retVal, err := satisfiesCondition(tctx.StartCondition(), cond, nil, tc.Input)
-
-			if tc.WantError {
-				require.Error(t, err)
-				return
-			}
-
-			require.NoError(t, err)
-			require.Equal(t, tc.Want, retVal)
-		})
-	}
-}
-
-func readCELTestCase(t *testing.T, data []byte) *privatev1.CelTestCase {
-	t.Helper()
-
-	tc := &privatev1.CelTestCase{}
-	require.NoError(t, util.ReadJSONOrYAML(bytes.NewReader(data), tc))
-
-	return tc
 }
 
 func readQPTestSuite(t *testing.T, data []byte) *privatev1.QueryPlannerTestSuite {
